requirements.txt
# Jarvis AI Assistant - Dependencies
# Python 3.10+ required

# AI & LLM
openai>=1.12.0              # DeepSeek API (OpenAI-compatible)

# Speech Recognition (Offline)
sounddevice>=0.4.6          # Audio input/output
soundfile>=0.12.1           # Audio file handling
soundfile>=0.12.1           # Audio file handling
SpeechRecognition>=3.10.0   # Google Speech API (Online)
vosk>=0.3.45                # Offline speech recognition (Backup)

# Text-to-Speech
edge-tts>=6.1.9             # Microsoft Edge TTS (free, high quality)

# System Automation
pyautogui>=0.9.54           # GUI automation
psutil>=5.9.8               # System monitoring
pyperclip>=1.8.2            # Clipboard operations
pynput>=1.7.6               # Global hotkeys

# Web & API
requests>=2.31.0            # HTTP requests
google-search-results>=2.4.2  # SerpAPI for web search

# GUI
Pillow>=10.2.0              # Image processing for UI

# Utilities
python-dotenv>=1.0.1        # Environment config

# Optional: Advanced features
# chromadb                  # For semantic memory search
# sentence-transformers     # For embeddings

# Mobile Remote
flask>=3.0.0
flask-cors>=4.0.0
